+++ cmd/odrive-database/migrations/1_idx_object_description.sql
@@ -1,0 +1,5 @@
+-- +migrate Up
+CREATE INDEX idx_object_description ON object (description);
+
+-- +migrate Down
+DROP INDEX `idx_object_description` ON object;